// src/classes/Entity/types.d.ts
export declare type PureAttributeDefinition = Partial<{
    partitionKey: false;
    sortKey: false;
    type: DynamoDBTypes;
    default: any | ((data: object) => any);
    dependsOn: string | string[];
    transform: (value: any, data: {}) => any;
    format: (value: any, data: {}) => any;
    coerce: boolean;
    save: boolean;
    onUpdate: boolean;
    hidden: boolean;
    required: boolean | 'always';
    alias: string;
    map: string;
    setType: DynamoDBKeyTypes;
    delimiter: string;
    prefix: string;
    suffix: string;
}>;